refactor: repackage properties
