--- src/main/java/com/programmers/vouchermanagement/configuration/properties/Customer.java
@@ -1,4 +1,0 @@
-package com.programmers.vouchermanagement.configuration.properties;
-
-public record Customer(String csvFileName, String jsonFileName) {
-}
