+++ src/main/java/com/programmers/vouchermanagement/configuration/properties/file/Customer.java
@@ -1,0 +1,4 @@
+package com.programmers.vouchermanagement.configuration.properties.file;
+
+public record Customer(String csvFileName, String jsonFileName) {
+}
