+++ src/test/java/com/programmers/vouchermanagement/customer/repository/FileCustomerRepositoryTest.java
@@ -22,7 +22,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 import com.programmers.vouchermanagement.configuration.TestConfig;
-import com.programmers.vouchermanagement.configuration.properties.FileProperties;
+import com.programmers.vouchermanagement.configuration.properties.file.FileProperties;
 import com.programmers.vouchermanagement.customer.domain.Customer;
 
 @SpringJUnitConfig(TestConfig.class)
