+++ src/test/java/com/programmers/vouchermanagement/voucher/repository/FileVoucherRepositoryTest.java
@@ -24,7 +24,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 import com.programmers.vouchermanagement.configuration.TestConfig;
-import com.programmers.vouchermanagement.configuration.properties.FileProperties;
+import com.programmers.vouchermanagement.configuration.properties.file.FileProperties;
 import com.programmers.vouchermanagement.voucher.domain.Voucher;
 import com.programmers.vouchermanagement.voucher.domain.VoucherType;
